Allow skipping of mtime in database store
The user may wish to completely disable mtime setting during the restore
of the database.  For doing that, she can use "git-meta set -M".
However, since the mtime key is always stored in the database, the
.gitmeta file may indicate, in some cases (for instance, when moving
files around without modifying them), modification times that are not
relevant.  This is annoying, because irrelevant changes keep getting
tracked uselessly by Git.

This commit avoids this, by preventing the mtime key to be stored when
option -M is used with "git-meta get".  The documentation string has
also been changed accordingly.

Note that it would also be interesting to have the same behavior for the
-U and -G options, but, in the current state of the code if neither -o
or -O options are given, the get_file_stats does not store uid and gid
information in the database, which amounts to skipping owner information
storage in "git-meta get".