Do not follow symlinks in getting/setting file attributes
This is done by using the argument follow_symlinks=False of functions
os.stat, os.utime, and os.chown.  The os.chmod function does not accept
that argument in Linux, simply because symlinks do not have
permissions.  Hence, we check whether the file is a symlink with
os.path.islink, before calling os.chmod.

Thess changes force the use of Python v3 and the code need to be
adjusted in several places:

1) One of the print statements needs to be changed by adding
parenthesis around its argument, because print is a function in Python 3.
2) The has_key() methods are replaced by the 'in' operator.
3) The string returned by subprocess.Popen.communicate is now a buffer
interface and needs to be decoded into UTF-8.