Sort entries in the JSON metadata file
Add the argument 'sort_keys=True' to the json.dump call, such that the
entries are sorted in the metadata file.  This seems to be necessary in
order to avoid gratuitous ordering changes in the metadata file, which
happen in successive calls "git meta get".

Also, this commits removes a useless statement "files.sort()".