Do not skip dangling symbolic links
It may happen that a symbolic link under Git control points to
nowhere.  In this case, the information about the symbolic link
must still be stored in the database.  Since os.path.isfile follows
symlinks, the test for the existence of symlink entries failed and
those entries were skipped in function git_get_files_attributes.
This commit fixes this problem.